refactor(scheduler): inline chunked-prefill packing into micro_scheduler.py
MicroScheduler already absorbs alternative selection strategies
(_select_node_priority, _select_node_rr) inline as helpers; the
chunked-step packing logic is the same shape and should live alongside
them rather than in its own module. Matches the codebase's convention of
keeping scheduler logic in one file.
